Homepage Updates
- Continued re-build of homepage category sliders to hardcoding.
Awaiting Images. Links active but most categories need reorganization on
the backend to display only the correct selections (i.e. The metal chess
sets category brings you to a submenu that allows you to select pieces
or full sets)
- Updated mah jong and backgammon store link to "carry over" the cart
between stores. Anything added to the cart in the chess store will be
carried over to either vstore upon click of the link.