Category page/homepage/Footer Updates
- Category page: Built category sidebar link navigation bar. This ONLY
includes the categories and dropdown menus for the categoies that are on
the current pages category sidebar. No links have been added at this
time pending listing and design by Jeff.
- Homepage: Corrected banner slider having an unnecessary margin on the
right side/white space and adjust image size to 1000px wide. Current
whitespace is due to it being present in image. Replacing with actual
images will remove it.
- Footer: Updated Read our Blog link with link value. Ready for testing.