Header/Product Page Updates
- Product: Corrected tabbing issue that prevented tabbing between
categories. Warranty and volume discount tab categories still being
displayed as there is no current backend data category for them to: 1.
Populate their data area with. 2. Determine if the tab should be visible
based on whether data is present or not.
- Product: Set you might like and accessories category displays to hide
if no products are in those categories.
- Product: Corrected Questions link. Previously it was not displaying or
linking correctly.
- Header: Corrected link issue with the see details link.